# swagger.yaml
openapi: 3.0.3
info:
  title: INTVoicekit API
  description: |
    Backend API for INTVoicekit - a voice assistant platform built on Express.js with MongoDB.
    This API acts as a proxy layer between clients and the external Vyom LiveKit API,
    managing local state for users, assistants, SIP trunks, tools, inbound configurations,
    and analytics.

    ## Authentication
    Every endpoint except `POST /api/auth/signup` and `POST /api/auth/login` requires the user's
    upstream LiveKit key, sent as `Authorization: Bearer <api_key>`. A missing or unknown key is
    a `401`. Obtain the key from signup or login. Identity comes from that key — a `user_id` in
    the request body or query string is ignored, and the key is forwarded to the external Vyom
    API on the user's behalf.

    ## Dual ID Resolution
    Many endpoints accept either a local MongoDB `_id` or an external service ID
    (e.g., `external_assistant_id`, `external_trunk_id`).

    ## External API Proxy
    Most service methods proxy requests to `https://api-livekit-vyom.indusnettechnologies.com/*`,
    forwarding the user's Bearer token. External responses are typically passed through directly.

    ## Assistant Runtime Modes
    `assistant_mode` selects the session shape. It is `pipeline` by default.
    - `pipeline` (half-cascade): a realtime model emits text, an external TTS provider speaks it.
      Vendor is `openai` only — `gemini` is rejected with `400` (see below). Requires
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/sip/create-outbound-trunk:
    post:
      tags: [SIP]
      summary: Create an outbound SIP trunk
      description: |
        Creates a SIP trunk both locally and externally. Supports Twilio and Exotel trunk types.
        Optionally enables passthrough mode for direct WebSocket connections.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trunk_name, trunk_type, trunk_config]
              properties:
                trunk_name:
                  type: string
                  example: "Production Twilio Trunk"
                trunk_type:
                  type: string
                  enum: [twilio, exotel]
                  example: "twilio"
                trunk_config:
                  type: object
                  description: |
                    For Twilio: { "address": "string", "numbers": ["string"], "username": "string", "password": "string" }
                    For Exotel: { "exotel_number": "string" }
                  properties:
                    address:
                      type: string
                      example: "pstn.twilio.com"
                    numbers:
                      type: array
                      items:
                        type: string
                      example: ["+1234567890"]
                    username:
                      type: string
                    password:
                      type: string
                    exotel_number:
                      type: string
                passthrough_mode:
                  type: boolean
                  default: false
                passthrough_webhook_url:
                  type: string
                  format: uri
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Outbound trunk created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Outbound trunk created successfully locally and externally"
                  trunk:
                    $ref: '#/components/schemas/SipTrunk'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sip/list:
    get:
      tags: [SIP]
      summary: List SIP trunks
      description: Retrieves all SIP trunks for a user from the local database.
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: SIP trunks retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SIP trunks retrieved successfully from local database"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SipTrunk'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sip/details/{id}:
    get:
      tags: [SIP]
      summary: Get SIP trunk details
      description: Retrieves details for a specific SIP trunk by local or external ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: SIP trunk details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/SipTrunk'
        '404':
          description: SIP trunk not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sip/delete/{id}:
    delete:
      tags: [SIP]
      summary: Delete a SIP trunk
      description: >
        Deactivates the trunk on the external API (`DELETE /sip/deactivate/{trunk_id}`, a
        soft delete that preserves the record for audit) and then removes the local mirror.

        If the external trunk is already gone (404) or already deactivated (400), the local
        row is still removed and `external_deactivated` comes back `false` — otherwise a
        half-finished delete would strand it locally forever. Any other upstream failure
        aborts before the local delete, so the two sides never drift apart.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: SIP trunk deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SIP trunk deactivated externally and removed locally"
                  external_deactivated:
                    type: boolean
                    description: >
                      `false` when the external trunk was already inactive or already gone.
                    example: true
                  data:
                    type: object
                    properties:
                      local_id:
                        type: string
                      trunk_id:
                        type: string
                  local_data_removed:
                    type: boolean
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== CALL ====================
...
  /api/tool/create:
    post:
      tags: [Tool]
      summary: Create a tool
      description: Creates a tool both locally and externally via the Vyom API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_name, tool_execution_type]
              properties:
                tool_name:
                  type: string
                  example: "Weather Check"
                tool_description:
                  type: string
                  example: "Fetches current weather for a location"
                tool_execution_type:
                  type: string
                  example: "webhook"
                tool_parameters:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      type:
                        type: string
                      required:
                        type: boolean
                  default: []
                tool_execution_config:
                  type: object
                  description: Configuration for tool execution (URL, headers, etc.)
                  properties:
                    url:
                      type: string
                      format: uri
                    method:
                      type: string
                      enum: [GET, POST, PUT, DELETE]
                    headers:
                      type: object
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Tool created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tool_id:
                        type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tool/list:
    get:
      tags: [Tool]
      summary: List tools
      description: Retrieves all tools from the external API.
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Tool list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tool/details/{id}:
    get:
      tags: [Tool]
      summary: Get tool details
      description: Retrieves details for a specific tool.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Tool details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tool/update/{id}:
    patch:
      tags: [Tool]
      summary: Update a tool
      description: Updates a tool on the external API.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tool_name:
                  type: string
                tool_description:
                  type: string
                tool_execution_type:
                  type: string
                tool_parameters:
                  type: array
                  items:
                    type: object
                tool_execution_config:
                  type: object
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Tool updated
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tool/delete/{id}:
    delete:
      tags: [Tool]
      summary: Delete a tool
      description: Deletes a tool from the external API.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Tool deleted
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tool/attach/{assistant_id}:
    post:
      tags: [Tool]
      summary: Attach tools to an assistant
      description: Attaches one or more tools to a specific assistant.
      parameters:
        - name: assistant_id
          in: path
          required: true
          schema:
            type: string
          description: Local _id or external_assistant_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_ids]
              properties:
                tool_ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  description: Array of local _id or external_tool_id
                  example: ["665f1a2b3c4d5e6f7a8b9c0d", "665f1a2b3c4d5e6f7a8b9c0e"]
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Tools attached
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tool/detach/{assistant_id}:
    post:
      tags: [Tool]
      summary: Detach tools from an assistant
      description: Detaches one or more tools from a specific assistant.
      parameters:
        - name: assistant_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_ids]
              properties:
                tool_ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Tools detached
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== WEB CALL ====================
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Something went wrong"
...
    SipTrunk:
      type: object
      description: >
        A trunk as returned by the list and details endpoints. `trunk_config` (which carries the
        Twilio `username` and `password`) is deliberately NOT returned — upstream withholds it
        from the list for security reasons and documents no details endpoint. It is sent only on
        create.
      properties:
        _id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        user_id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        external_trunk_id:
          type: string
          example: "ext_trunk_xyz789"
        trunk_name:
          type: string
          example: "Production Twilio Trunk"
        trunk_type:
          type: string
          enum: [twilio, exotel]
          example: "twilio"
        passthrough_mode:
          type: boolean
          example: false
        passthrough_webhook_url:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The user's upstream LiveKit `api_key`, obtained from `POST /api/auth/signup` or
        `POST /api/auth/login`, sent as `Authorization: Bearer <api_key>`.

# Applied to every operation; /api/auth/signup and /api/auth/login override it with `security: []`.
security:
  - bearerAuth: []